refactor(marazion): authored/primitive split; drop complex graph
`HydrologyComplex` is now a bag of blendable `HydrologyNode`s (no
`WatershedEdge` / hub graph / `wet_union` / `depression`). Module tree
splits leaf authoring from runtime primitives; hydro elevation/footprint
variants are proper structs under submodules (no `mod.rs`).

## Agent Dialogue

- **Developer:** Clean up outdated concepts; split authored vs primitive; struct-ify HydroElevation variants; drop complex edges (complex = nodes to blend). No `mod.rs`.
- **Agent:** Removed graph/depression/wet_union; reorganized into `authored/` + `primitive/`; `HydroElevation::{Reach,Radial}` and `HydroFootprint::{Reach,Ellipse}` wrap dedicated structs; renamed `occupancy_at` → `min_phi_at`.
