+++ maybraid/durham/marazion/src/authored.rs
@@ -1,0 +1,11 @@
+//! Leaf / graph authoring plans that emit [`crate::primitive::node::HydrologyNode`]s.
+
+pub mod apron;
+pub mod bog;
+pub mod lake;
+pub mod noise;
+pub mod pocket_cell;
+pub mod polyline;
+pub mod pre_pocket;
+pub mod stream;
+pub mod streams_graph;
