+++ maybraid/durham/marazion/src/authored/pocket_cell.rs
@@ -1,6 +1,6 @@
 //! Pocket-cell guillotine partition — [RFC-127 §3.1.2](https://github.com/ramate-io/maybraid/tree/main/rfc/rfc-000-000-127-marazion-watersheds#312-pocket-cells).
 
-use crate::noise::n01;
+use crate::authored::noise::n01;
 use bevy_math::Vec2;
 use procedural_common::Bounds2;
