+++ maybraid/durham/marazion/src/primitive/complex.rs
@@ -1,47 +1,16 @@
 //! Indexed hydrology complex: member nodes + sample-time correction.
 //!
-//! Authored leaves emit [`crate::node::HydrologyNode`]s. [`HydrologyComplex`]
-//! indexes them and blends terrain via
-//! [`HydrologyNode::blend_terrain_elevation`]. Optional backfills stay a
-//! post-depression jersey layer via [`Self::compile`].
+//! Authored leaves emit [`crate::primitive::node::HydrologyNode`]s.
+//! [`HydrologyComplex`] is a bag of nodes to blend — no graph / connectivity.
+//! Optional backfills stay a post-depression jersey layer via [`Self::compile`].
 
-use crate::backfill::WatershedBackfill;
-use crate::depression::WatershedDepression;
-use crate::fill::{WaterFill, WaterSurface};
-use crate::hydro::{ComplexApronParams, FootprintIndex, HydroPrimitive};
-use crate::node::{HydroParameters, HydrologyNode};
+use crate::primitive::backfill::WatershedBackfill;
+use crate::primitive::fill::{WaterFill, WaterSurface};
+use crate::primitive::hydro::{ComplexApronParams, FootprintIndex, HydroPrimitive};
+use crate::primitive::node::{HydroParameters, HydrologyNode};
 use bevy_math::{Vec2, Vec3};
-use jersey_terrain_stamps::{JerseyModulation, Region2D};
+use jersey_terrain_stamps::JerseyModulation;
 use procedural_common::Bounds2;
-
-/// Stable index into [`HydrologyComplex::graph_nodes`].
-pub type WatershedNodeId = usize;
-/// Stable index into [`HydrologyComplex::edges`].
-pub type WatershedEdgeId = usize;
-
-/// Graph node: empty hub, joint polyline, or lake bowl depression.
-#[derive(Debug, Clone)]
-pub struct WatershedNode {
-	pub depression: Option<WatershedDepression>,
-}
-
-impl WatershedNode {
-	pub fn empty() -> Self {
-		Self { depression: None }
-	}
-
-	pub fn with_depression(depression: WatershedDepression) -> Self {
-		Self { depression: Some(depression) }
-	}
-}
-
-/// Graph edge: typically a stream-corridor depression between two nodes.
-#[derive(Debug, Clone)]
-pub struct WatershedEdge {
-	pub from: WatershedNodeId,
-	pub to: WatershedNodeId,
-	pub depression: WatershedDepression,
-}
 
 /// Stamp products derived from a [`HydrologyComplex`] (fills + optional backfills).
 #[derive(Debug, Clone)]
@@ -53,7 +22,6 @@ pub struct CompiledWatershed {
 	/// Post-depression jersey ops (backfills).
 	pub modulations: Vec<JerseyModulation>,
 	pub fills: Vec<WaterFill>,
-	pub wet_union: Option<Region2D>,
 }
 
 impl CompiledWatershed {
@@ -80,8 +48,6 @@ impl CompiledWatershed {
 pub struct HydrologyComplex {
 	pub bounds: Bounds2,
 	pub seed: u32,
-	pub graph_nodes: Vec<WatershedNode>,
-	pub edges: Vec<WatershedEdge>,
 	pub backfills: Vec<WatershedBackfill>,
 	/// Member hydrology nodes (source of truth for correction).
 	pub hydrology: Vec<HydrologyNode>,
@@ -94,8 +60,6 @@ impl HydrologyComplex {
 		let mut complex = Self {
 			bounds,
 			seed,
-			graph_nodes: Vec::new(),
-			edges: Vec::new(),
 			backfills: Vec::new(),
 			hydrology: Vec::new(),
 			index: FootprintIndex::empty(),
@@ -139,18 +103,6 @@ impl HydrologyComplex {
 		self
 	}
 
-	pub fn push_node(&mut self, node: WatershedNode) -> WatershedNodeId {
-		let id = self.graph_nodes.len();
-		self.graph_nodes.push(node);
-		id
-	}
-
-	pub fn push_edge(&mut self, edge: WatershedEdge) -> WatershedEdgeId {
-		let id = self.edges.len();
-		self.edges.push(edge);
-		id
-	}
-
 	pub fn is_empty(&self) -> bool {
 		self.hydrology.is_empty()
 	}
@@ -159,23 +111,6 @@ impl HydrologyComplex {
 		let short = self.bounds.extent().min_element().max(1.0);
 		let cell = (short * 0.08).clamp(8.0, 64.0);
 		self.index = FootprintIndex::build_nodes(self.bounds, &self.hydrology, cell);
-	}
-
-	fn wet_union_from_graph(&self) -> Option<Region2D> {
-		let mut wet_cores: Vec<Region2D> = Vec::new();
-		for node in &self.graph_nodes {
-			if let Some(dep) = &node.depression {
-				wet_cores.push(dep.wet_core.clone());
-			}
-		}
-		for edge in &self.edges {
-			wet_cores.push(edge.depression.wet_core.clone());
-		}
-		match wet_cores.len() {
-			0 => None,
-			1 => wet_cores.pop(),
-			_ => Some(Region2D::union(wet_cores)),
-		}
 	}
 
 	pub fn candidate_ids(&self, p: Vec2) -> Vec<u16> {
@@ -221,8 +156,8 @@ impl HydrologyComplex {
 		HydrologyNode::blend_surface_elevation(&nodes, p)
 	}
 
-	/// Min occupancy \(\phi\) over intersecting nodes (`None` if none in index).
-	pub fn occupancy_at(&self, x: f32, z: f32) -> Option<f32> {
+	/// Min \(\phi\) over intersecting nodes (`None` if none in index).
+	pub fn min_phi_at(&self, x: f32, z: f32) -> Option<f32> {
 		let p = Vec2::new(x, z);
 		let mut phi = f32::INFINITY;
 		for node in self.nodes_intersecting(p) {
@@ -238,10 +173,10 @@ impl HydrologyComplex {
 
 	/// Cheap exterior distance toward the wet carve (\(\phi = 0\)).
 	///
-	/// Prefer indexed occupancy; if the sample misses the broadphase, fall back to
+	/// Prefer indexed \(\phi\); if the sample misses the broadphase, fall back to
 	/// a full member scan so MC sees a continuous field instead of a huge sentinel.
 	pub fn approximate_distance_to_carve(&self, x: f32, z: f32) -> f32 {
-		if let Some(phi) = self.occupancy_at(x, z) {
+		if let Some(phi) = self.min_phi_at(x, z) {
 			return phi;
 		}
 		let p = Vec2::new(x, z);
@@ -279,17 +214,23 @@ impl HydrologyComplex {
 	}
 
 	pub fn compile(&self) -> CompiledWatershed {
-		let wet_union = self.wet_union_from_graph();
-		let modulations: Vec<_> =
-			self.backfills.iter().cloned().map(|b| b.into_modulation()).collect();
-		let fills = if self.is_empty() { Vec::new() } else { vec![self.water_fill()] };
+		let modulations: Vec<_> = self
+			.backfills
+			.iter()
+			.cloned()
+			.map(|b| b.into_modulation())
+			.collect();
+		let fills = if self.is_empty() {
+			Vec::new()
+		} else {
+			vec![self.water_fill()]
+		};
 		CompiledWatershed {
 			bounds: self.bounds,
 			seed: self.seed,
 			complex: self.clone(),
 			modulations,
 			fills,
-			wet_union,
 		}
 	}
 }
@@ -297,10 +238,12 @@ impl HydrologyComplex {
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use crate::backfill::WatershedBackfill;
-	use crate::depression::{WatershedDepression, WatershedDepressionKind};
-	use crate::hydro::{HydroElevation, HydroFootprint, HydroPrimitive};
-	use crate::node::{HydroParameters, HydrologyNode};
+	use crate::primitive::backfill::WatershedBackfill;
+	use crate::primitive::hydro::{
+		Ellipse, HydroElevation, HydroFootprint, HydroPrimitive, RadialBowl, ReachProfile,
+		ReachSegment,
+	};
+	use crate::primitive::node::{HydroParameters, HydrologyNode};
 	use jersey_terrain_stamps::{CircleRegion, Region2D, RegionNoise};
 
 	#[test]
@@ -314,26 +257,27 @@ mod tests {
 
 	#[test]
 	fn backfill_appends_after_hydro() -> anyhow::Result<()> {
-		let core = Region2D::Circle(CircleRegion { center: Vec2::ZERO, radius: 10.0 });
+		let core = Region2D::Circle(CircleRegion {
+			center: Vec2::ZERO,
+			radius: 10.0,
+		});
 		let node = HydrologyNode::new(
 			HydroPrimitive {
-				footprint: HydroFootprint::Ellipse {
+				footprint: HydroFootprint::Ellipse(Ellipse {
 					center: Vec2::ZERO,
 					radii: Vec2::splat(8.0),
 					rotation: 0.0,
-				},
-				elevation: HydroElevation::RadialBowl { surface: 40.0, center_depth: 3.0 },
+				}),
+				elevation: HydroElevation::Radial(RadialBowl {
+					surface: 40.0,
+					center_depth: 3.0,
+				}),
 				influence_pad: 4.0,
 			},
 			HydroParameters::default(),
 			12.0,
 		);
-		let mut complex = HydrologyComplex::new(Bounds2::from_xz(-40.0, -40.0, 40.0, 40.0), 3);
-		complex.push_node(WatershedNode::with_depression(WatershedDepression::new(
-			WatershedDepressionKind::LakeBowl,
-			core.clone(),
-		)));
-		let out = complex
+		let out = HydrologyComplex::new(Bounds2::from_xz(-40.0, -40.0, 40.0, 40.0), 3)
 			.with_hydrology(vec![node])
 			.with_backfill(WatershedBackfill::basin(
 				core,
@@ -343,26 +287,6 @@ mod tests {
 			.compile();
 		assert!(out.has_hydro());
 		assert_eq!(out.modulations.len(), 1);
-		Ok(())
-	}
-
-	#[test]
-	fn wet_union_of_two_cores_is_min_sdf() -> anyhow::Result<()> {
-		let a = Region2D::Circle(CircleRegion { center: Vec2::new(-10.0, 0.0), radius: 4.0 });
-		let b = Region2D::Circle(CircleRegion { center: Vec2::new(10.0, 0.0), radius: 4.0 });
-		let mut complex = HydrologyComplex::new(Bounds2::from_xz(-40.0, -40.0, 40.0, 40.0), 2);
-		complex.push_node(WatershedNode::with_depression(WatershedDepression::new(
-			WatershedDepressionKind::LakeBowl,
-			a.clone(),
-		)));
-		complex.push_node(WatershedNode::with_depression(WatershedDepression::new(
-			WatershedDepressionKind::LakeBowl,
-			b.clone(),
-		)));
-		let out = complex.compile();
-		let u = out.wet_union.expect("union");
-		let p = Vec2::new(-10.0, 0.0);
-		assert!((u.sdf(p) - a.sdf(p).min(b.sdf(p))).abs() < 1e-5);
 		Ok(())
 	}
 
@@ -383,16 +307,16 @@ mod tests {
 		let c = Vec2::new(80.0, 0.0);
 		let upstream = HydrologyNode::new(
 			HydroPrimitive {
-				footprint: HydroFootprint::ReachSegment {
+				footprint: HydroFootprint::Reach(ReachSegment {
 					a,
 					b,
 					half_width: half_w,
-				},
-				elevation: HydroElevation::ReachProfile {
+				}),
+				elevation: HydroElevation::Reach(ReachProfile {
 					surface_a: 100.0,
 					surface_b: 80.0,
 					center_depth: depth,
-				},
+				}),
 				influence_pad: rim_w + 8.0,
 			},
 			params.clone(),
@@ -400,16 +324,16 @@ mod tests {
 		);
 		let downstream = HydrologyNode::new(
 			HydroPrimitive {
-				footprint: HydroFootprint::ReachSegment {
+				footprint: HydroFootprint::Reach(ReachSegment {
 					a: b,
 					b: c,
 					half_width: half_w,
-				},
-				elevation: HydroElevation::ReachProfile {
+				}),
+				elevation: HydroElevation::Reach(ReachProfile {
 					surface_a: 80.0,
 					surface_b: 60.0,
 					center_depth: depth,
-				},
+				}),
 				influence_pad: rim_w + 8.0,
 			},
 			params,
@@ -438,15 +362,15 @@ mod tests {
 	fn water_exterior_uses_carve_distance_not_sentinel() -> anyhow::Result<()> {
 		let node = HydrologyNode::new(
 			HydroPrimitive {
-				footprint: HydroFootprint::Ellipse {
+				footprint: HydroFootprint::Ellipse(Ellipse {
 					center: Vec2::ZERO,
 					radii: Vec2::splat(10.0),
 					rotation: 0.0,
-				},
-				elevation: HydroElevation::RadialBowl {
+				}),
+				elevation: HydroElevation::Radial(RadialBowl {
 					surface: 40.0,
 					center_depth: 3.0,
-				},
+				}),
 				influence_pad: 20.0,
 			},
 			HydroParameters::default(),
@@ -471,32 +395,28 @@ mod tests {
 	fn water_fill_probes_node_interior() -> anyhow::Result<()> {
 		let node = HydrologyNode::new(
 			HydroPrimitive {
-				footprint: HydroFootprint::Ellipse {
+				footprint: HydroFootprint::Ellipse(Ellipse {
 					center: Vec2::new(30.0, -25.0),
 					radii: Vec2::splat(8.0),
 					rotation: 0.0,
-				},
-				elevation: HydroElevation::RadialBowl {
+				}),
+				elevation: HydroElevation::Radial(RadialBowl {
 					surface: 40.0,
 					center_depth: 3.0,
-				},
+				}),
 				influence_pad: 12.0,
 			},
 			HydroParameters::default(),
 			12.0,
 		);
-		let mut complex = HydrologyComplex::new(Bounds2::from_xz(-40.0, -40.0, 40.0, 40.0), 5);
-		complex.push_node(WatershedNode::with_depression(WatershedDepression::new(
-			WatershedDepressionKind::LakeBowl,
-			Region2D::Circle(CircleRegion {
-				center: Vec2::new(30.0, -25.0),
-				radius: 8.0,
-			}),
-		)));
-		let fill = complex.with_hydrology(vec![node]).water_fill();
+		let fill = HydrologyComplex::new(Bounds2::from_xz(-40.0, -40.0, 40.0, 40.0), 5)
+			.with_hydrology(vec![node])
+			.water_fill();
 		let probes = fill.wet_volume_probe_points();
 		assert!(
-			probes.iter().any(|p| (p.x - 30.0).abs() < 1e-3 && (p.y + 25.0).abs() < 1e-3),
+			probes
+				.iter()
+				.any(|p| (p.x - 30.0).abs() < 1e-3 && (p.y + 25.0).abs() < 1e-3),
 			"expected lake-center probe, got {probes:?}"
 		);
 		Ok(())
